% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pols_501.R
\name{coalesce}
\alias{coalesce}
\title{First non-missing element for vectors}
\usage{
coalesce(...)
}
\arguments{
\item{...}{vectors}
}
\value{
For each element, the value of the first non \code{NA} element in the
vectors of \code{...}
}
\description{
First non-missing element for vectors
}
\author{
\href{http://stackoverflow.com/users/903061/gregor}{Gregor}, \href{http://stackoverflow.com/users/2588184/mrip}{mrip}
}
\references{
\url{http://stackoverflow.com/questions/19253820/}
}

