% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pols_501.R
\name{extract_var_info}
\alias{extract_var_info}
\title{Pulls label names from STATA datasets imported to R}
\usage{
extract_var_info(dataset)
}
\arguments{
\item{dataset}{An dataframe containing a STATA dataset
imported to R}
}
\description{
Pulls label names from STATA datasets imported to R
}

