% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pols_501.R
\name{fill_na}
\alias{fill_na}
\title{Fill missing values}
\usage{
fill_na(x, fill, ...)
}
\arguments{
\item{x}{vector}

\item{fill}{Value to replace missing values with.}

\item{...}{other parameters}
}
\description{
Fill missing values
}

