% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pols_501.R
\name{filter_na}
\alias{filter_na}
\alias{filter_na_}
\title{Filter rows with missing values}
\usage{
filter_na(.data, ...)

filter_na_(.data, ..., .dots)
}
\arguments{
\item{.data}{A tbl. Currently only methods for \code{\link{data.frame}} and
\code{\link{tbl_df}} are provided.}

\item{...}{Comma separated list of unquoted expressions to select variables.
See \code{\link[dplyr]{select}}.}

\item{.dots}{Use \code{filter_na_()} to do standard evaluation. See
\code{vignette("nse", package = "dplyr")} for details.}
}
\value{
An object of the same class as tbl.
}
\description{
\code{filter_na()} filters rows with missing (\code{NA}) values.
}

