% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pols_501.R
\name{is_same}
\alias{is_same}
\title{Compare vectors with missing values}
\usage{
is_same(x, y)
}
\arguments{
\item{x, y}{vectors to compare}
}
\value{
A logical vector: \code{TRUE} if both vectors have the same value, including
\code{NA}, and \code{FALSE}, if they differ.
}
\description{
Compare two vectors treating \code{NA} like any other value.
If both vectors have an \code{NA} for that element, the function returns \code{TRUE},
if one has an \code{NA} and the other does not, it returns \code{FALSE}.
}
\author{
Winston Chang
}
\references{
\url{http://www.cookbook-r.com/Manipulating_data/Comparing_vectors_or_factors_with_NA/}
}

