% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iver}
\alias{iver}
\title{Party Systems and Redistributive Government}
\format{A data frame containng 4 variables and 14 observations.

\describe{
 \item{cty}{Country}
 \item{elec_sys}{Electoral system: Majoritarian "maj", Proportional representation "pr", unamimity government "unam"}
 \item{povred}{Percent of citizens liftedn out of poverty by taxes and transfers.}
 \item{enp}{Effective number of parties.}
}}
\usage{
iver
}
\description{
Party Systems and Redistributive Government
}
\references{
Torben Iversen and David Soskice. 2006. ``Electoral Institutions and the Politics of Coalitions: Why Some
 Democracies Redistribute More Than Others.`` \emph{American Political Science Review}.
}
\keyword{datasets}

