% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pols_501.R
\name{naif}
\alias{naif}
\title{Return NA if x == y, and x otherwise}
\usage{
naif(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
vector, with \code{NA} where \code{x == y}, and \code{x} otherwise.
}
\description{
Return NA if x == y, and x otherwise
}
\details{
This is similar to the SQL function \code{NULLIF}. It can be used
to peform the inverse operation of \code{\link{coalesce}}.
}

