% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Prestige}
\alias{Prestige}
\title{Prestige of Canadian Occupations}
\format{The Prestige data frame has 102 rows and 6 columns. The observations are occupations.

\describe{
 \item{education}{Average education of occupational incumbents, years, in 1971.}
 \item{income}{Average income of incumbents, dollars, in 1971.}
 \item{women}{Percentage of incumbents who are women.}
 \item{prestige}{Pineo-Porter prestige score for occupation, from a social survey conducted in the mid-1960s.}
 \item{census}{Canadian Census occupational code.}
 \item{type}{Type of occupation. A factor with levels (note: out of order): bc, Blue Collar; prof, Professional, Managerial, and Technical; wc, White Collar.}
}}
\source{
car package
}
\usage{
Prestige
}
\description{
Prestige of Canadian Occupations
}
\keyword{datasets}

