% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pols_501.R
\docType{import}
\name{reexports}
\alias{mapvalues}
\alias{reexports}
\alias{revalue}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{plyr}{\code{\link[plyr]{revalue}}, \code{\link[plyr]{mapvalues}}}
}}
\keyword{internal}

