% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pols_501.R
\name{replace_na}
\alias{replace_na}
\title{Replace values with NA}
\usage{
replace_na(x, values = NULL)
}
\arguments{
\item{x}{vector}

\item{values}{Values which should be replaced by missing values}
}
\description{
Replace values with NA
}

