% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turnout}
\alias{turnout}
\title{Turnout Data Set from the 1992 American National Election Survey}
\format{A table containing 5 variables ("race", "age", "educate",  "income", and "vote") and 2,000 observations.}
\source{
Zelig package
}
\usage{
turnout
}
\description{
This data set contains individual-level turnout data. It pools several
American National Election Surveys conducted during the 1992 presidential
election year.  Only the first 2,000 observations (from a total of 15,837
                                                    observations) are included in the sample data.
}
\references{
King, Gary, Michael Tomz, Jason Wittenberg (2000).
  ``Making the Most of Statistical Analyses: Improving Interpretation and
  Presentation,'' \emph{American Journal of Political Science}, vol. 44,
  pp.341--355.
}
\keyword{datasets}

