% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pols_501.R
\name{zip_project}
\alias{zip_project}
\title{Zip project or assignment for submission}
\usage{
zip_project(use_tar = FALSE)
}
\arguments{
\item{use_tar}{Create a tar.gz file instead of a zip file.}
}
\description{
Will create a zipfile of the current directory that you can submit as an assignment.
}

