% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pols_501.R
\name{create_assignment_skeleton}
\alias{create_assignment_skeleton}
\title{Create Directory Structure for an Assignment}
\usage{
create_assignment_skeleton(assignment, netId, dir = ".")
}
\arguments{
\item{assignment}{Assignment number}

\item{netId}{Student netID. Example: jarnold}

\item{dir}{Directory in which the assignment will be created}
}
\description{
Create Directory Structure for an Assignment
}

